+++ src/NEventStore.Serialization.Bson/BsonSerializationWireupExtension.cs
@@ -1,0 +1,10 @@
+﻿namespace NEventStore.Serialization.Bson
+{
+    public static class BsonSerializationWireupExtension
+    {
+        public static SerializationWireup UsingBsonSerialization(this PersistenceWireup wireup)
+        {
+            return wireup.UsingCustomSerialization(new BsonSerializer());
+        }
+    }
+}