+++ src/NEventStore.Serialization.Json/JsonSerializer.cs
@@ -50,28 +50,31 @@ namespace NEventStore.Serialization.Json
         {
             if (Logger.IsVerboseEnabled) Logger.Verbose(Messages.SerializingGraph, typeof(T));
             using (var streamWriter = new StreamWriter(output, Encoding.UTF8))
-                Serialize(new JsonTextWriter(streamWriter), graph);
+            using (var jsonTextWriter = new JsonTextWriter(streamWriter))
+            {
+                Serialize(jsonTextWriter, graph);
+            }
         }
 
         public virtual T Deserialize<T>(Stream input)
         {
             if (Logger.IsVerboseEnabled) Logger.Verbose(Messages.DeserializingStream, typeof(T));
             using (var streamReader = new StreamReader(input, Encoding.UTF8))
-                return Deserialize<T>(new JsonTextReader(streamReader));
+            using (var jsonTextReader = new JsonTextReader(streamReader))
+            {
+                return Deserialize<T>(jsonTextReader);
+            }
         }
 
         protected virtual void Serialize(JsonWriter writer, object graph)
         {
-            using (writer)
-                GetSerializer(graph.GetType()).Serialize(writer, graph);
+            GetSerializer(graph.GetType()).Serialize(writer, graph);
         }
 
         protected virtual T Deserialize<T>(JsonReader reader)
         {
             Type type = typeof(T);
-
-            using (reader)
-                return (T)GetSerializer(type).Deserialize(reader, type);
+            return (T)GetSerializer(type).Deserialize(reader, type);
         }
 
         protected virtual Newtonsoft.Json.JsonSerializer GetSerializer(Type typeToSerialize)